Move data source config classes from pipeline-core module to pipeline-spi module
